requirements.txt
streamlit
numpy
scikit-learn
nltk
pillow
torch
torchvision
grad-cam